requirements.txt
discord.py>=1.4.1
ruamel.yaml>=0.16.12
